Make horizon dockerfile and update latest.sql

+++ services/horizon/internal/db2/schema/latest.sql
@@ -319,7 +319,7 @@ CREATE TABLE history_transactions (
     application_order integer NOT NULL,
     account character varying(64) NOT NULL,
     account_sequence bigint NOT NULL,
-    fee_paid integer NOT NULL,
+    fee_paid bigint NOT NULL,
     operation_count integer NOT NULL,
     created_at timestamp without time zone,
     updated_at timestamp without time zone,
